afsocket-dest: do not lose messages when DNS resolving fails
The problem was that the queue, which stores the incoming messages,
wasn't set for the afsocket destination when DNS resolving failed.

Technical details:

1. setup LogWriter instance
Queues are bound to LogWriter instances. For afsocket destination,
LogWriter instance is setup only after the network addresses are
successfully setup (`afsocket_dd_try_connect`).
This means that if DNS resolving fails then LogWriter instance
isn't setup for the destination.

When we have an initialized LogWriter instance then we shouldn't re-setup it.
This is why we check it in the `afsocket_dd_setup_writer`.

So the first problem what we have to solve is to setup a LogWriter
instance even when the `afsocket_dd_setup_addresses` fails.
TODO: when this function fails??? is this new behaviour correct for
every cases?

2. do not leak an existing LogWriter instance
When the afsocket-dest had a LogWriter instance then we lost it
when we called `afsocket_dd_restore_connection` because of the changes
initiated in 1)
Why?
When we call `afsocket_dd_setup_connection` then we call
`afsocket_dd_restore_connection` which just fetch LogWriter instance
from the persist_config. But persist_config doesn't contain such an instance
as we didn't call the deinit, so we got a NULL pointer what we set as
our LogWriter instance and we leak the original one (and lose all the messages in
the attached queue...)
This is why I try to restore the LogWriter instance only when it is NULL.
(and it is NULL if `afsocket_dd_deinit` was called because that function
calls `afsocket_dd_save_connection` if the connection has been initialized
which is true if we successfully setup the LogWriter instance).

3. persist name
When we acquire the queue for the LogWriter instance we use the persist name.
This name contains the `dest_name` which contains the address information (host
and port). The port is retrieved from a GSockAddr instance which is
created by the `afsocket_dd_setup_addresses`. But we know that this function
failed in 1) which means that this instance never created: one way ticket to
a nice segfault... This is why I use the `dest_port` in
`afinet_dd_get_dest_name`.

Setup LogWriter instance during init phase.
When log_writer_init fails then init is also fails.

TODO: fix commit message
split into smaller patches : deterimine_port, etc...

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>
Signed-off-by: Laszlo Budai <stentor.bgyk@gmail.com>