add-contextual-data: selector factored out as a separate class
Concept:
Previously only one kind of selector was available: template based.
When a LogMessage passed to the add-contextual-data parser, the result
string was generated by formatting the template for the incoming message.
This string was used as a key for ContextInfoDb and the corresponding
name-value pair was appended to the message.

With this change, the Selector has been transormed to a generalized interface
and the template-based selector is the first implementation of it.
The main method of this interface is `gchar* resolve(LogMessage *)`.
In this way it is possible to implement more sophisticated Selectors, not only
template based and these Selectors can be implemented as plugins in the future.

Signed-off-by: Laszlo Budai <Laszlo.Budai@balabit.com>