secret_storage: test_rlimit unit test removed
Unit test test_rlimit is removed because it cannot work if executed by
root. The test is intended to check how MEMLOCK_RLIMIT affects secret
storage. However if executed by root, rlimits are ignored. Note
dropping CAP_SYS_RESOURCE cannot help, because it only controls if a
process can call setrlimit or not. It does not control if a process
can ignore the rlimits.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>