logsource: memory-limited-flow-control: add memory-limit(guint64) option
Memory limited flow control: iw-size, log-fifo-size could be replaced by a
memory limit. When a message is received, the current memory counter is
decreased by the processed message size (the msg size after it passed
through the processing pipeline, so this is the final size of the msg).
When the message is acked back and freed, the counter is decremented.

When the counter reach the limit, the source become suspended. New
messages will be received only after the counter is below under the
limit.

The memory limit is defined per source.

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>