logqueue-fifo: Drop messages only when flow-control is disabled.
After flow control is based on the memory usage of LogSources,
in the flow-controlled case we can keep these messages: they won't use more
memory... and the source is already in suspended state, so it's better
to keep these messages.

In the non-flow-controlled case we have to drop them to keep memory
usage below the limit.

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>