wildcard-file: use AckTracker::on_all_acked to free reader
The best would be if we could merge _remove_file_reader() and
_stop_file_reader(). This way, the deinitialization of the
wildcard reader would be the same as what we have in afsocket.

Unfortunately, this is not possible without changing a few things around
bookmark handling, because:

- When a file is deleted, its old persist state has to be removed +
  the reader needs to be removed from the hash table.
- A new file with the same name can be registered, while syslog-ng is
  reading the old "deleted" one.
  In this case, the new file will be delayed in order to avoid persist name
  collision. We also have to wait all the acks of the old file before
  starting the new one.
  This is why "last_msg_sent" is needed.

A better implementation would be if we could instruct the bookmark handler
logic (in general or in _buffered_server_bookmark_save()) to stop
persisting its bookmarks when the file is deleted, so we should not wait
for the old file to be finished.

Signed-off-by: László Várady <laszlo.varady@balabit.com>