logsource: delay setting the default value of init_window_size
In some cases, the default value of init_window_size should depend on
other config options (for example: dynamic window).

Since this type of information (whether the dependent option was set or
not) is available after parsing the config file completely, so the default
value has to be set in log_source_options_init().

Signed-off-by: László Várady <laszlo.varady@balabit.com>